+++ internal/service/commit/server.go
@@ -14,7 +14,3 @@ var defaultBranchName = ref.DefaultBranchName
 func NewServer() pb.CommitServiceServer {
 	return &server{}
 }
-
-func (s *server) FindCommits(in *pb.FindCommitsRequest, stream pb.CommitService_FindCommitsServer) error {
-	return nil
-}