+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Gitaly changelog
 
+UNRELEASED
+
+- Implement CommitService.FindCommits
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/266
+
 v0.30.0
 
 - Add a middleware for handling Git object dir attributes